1) разделены по папкам файлы согласно комментарию: main - в основной, каждый класс в отдельном файле, для каждого класса отдельный тест 2) поправила атрибуты класса в Category 3) исправила тесты в Category
